requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0
trl>=0.7.0
wandb>=0.15.0
tensorboard>=2.13.0
pyyaml>=6.0
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
jieba>=0.42.1
pydantic>=2.0.0
loguru>=0.7.0
huggingface_hub>=0.19.0 